Standardize SCMProbe usage; add a bit more useful logs
